Measure a design unit with a float, as every other length is
They were integers because that is how the font file stores most of
them, but nothing downstream ever used the integer-ness: every one is
either compared against another design unit or turned into a float at
once. The conversion that did that, and the auto-opened module holding
it, are both gone.

What is left counting with an integer is what genuinely counts: a glyph
id, a codepoint, an index, the number of assembly parts, and TeX's
eighteenths of an em.

No render changed.

+++ ModeMath/MathFontTypes.fs
@@ -13,21 +13,21 @@ type Glyph
     (
         face: Face,
         id: int,
-        advance: int<du>,
-        top: int<du>,
-        bottom: int<du>,
-        italicCorrection: int<du>,
-        topAccentAttachment: int<du>
+        advance: float32<du>,
+        top: float32<du>,
+        bottom: float32<du>,
+        italicCorrection: float32<du>,
+        topAccentAttachment: float32<du>
     ) =
     /// A glyph of the math face, which everything but blackboard bold is set in.
     internal new
         (
             id: int,
-            advance: int<du>,
-            top: int<du>,
-            bottom: int<du>,
-            italicCorrection: int<du>,
-            topAccentAttachment: int<du>
+            advance: float32<du>,
+            top: float32<du>,
+            bottom: float32<du>,
+            italicCorrection: float32<du>,
+            topAccentAttachment: float32<du>
         ) =
         Glyph(Face.Math, id, advance, top, bottom, italicCorrection, topAccentAttachment)
     /// The file the id indexes, which the painter needs to draw it.
@@ -54,7 +54,7 @@ type CG internal (codepoint: int, glyph: Glyph) =
 
 /// One ready-made size of a glyph that stretches.
 [<Struct>]
-type StretchSize internal (glyph: Glyph, advance: int<du>) =
+type StretchSize internal (glyph: Glyph, advance: float32<du>) =
     member _.Glyph = glyph
     /// The size to choose on: a height along the vertical axis, a width along the horizontal.
     member _.Advance = advance
@@ -65,9 +65,9 @@ type AssemblyPart
     internal
     (
         glyph: Glyph,
-        startConnector: int<du>,
-        endConnector: int<du>,
-        fullAdvance: int<du>,
+        startConnector: float32<du>,
+        endConnector: float32<du>,
+        fullAdvance: float32<du>,
         isExtender: bool
     ) =
     member _.Glyph = glyph
